+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Don't pass sleep assessment link in raw response(This is to avoid GPT generating sleep assessment in symptom queries)
 - Use search product for symptom related queries(This will suggest different products for insomnia and sleep apnea)
+- Formatted product response 
+- Don't include "We recommend you take an assessment and also speak to a Doctor." in raw_response(message_log). Because it gets repeated in future when user ask query which suggest assessment
 
 ### Changed
 - Updated SYSTEM_PROMPT